[seller] refactor(173): store와 관계된 type을 entity로 이동, 주석 및 import 정리
