+++ apps/seller/src/features/products/create/create-form/index.ts
@@ -3,4 +3,3 @@ export { InfoTooltip } from './info-tooltip.ui'
 // export { useCreateFormSteps } from './use-create-form-steps.hook'
 export { useCreateProductForm } from './use-create-product-form.hook'
 export { ProductFinalPrice } from './product-final-price.ui'
-export * from './product-create.types'
